fix: update tool path
