--- fixtures/servers/calc/index.js
@@ -1,75 +1,0 @@
-import { McpServer } from '@modelcontextprotocol/sdk/server/mcp.js';
-import { StdioServerTransport } from '@modelcontextprotocol/sdk/server/stdio.js';
-import { z } from 'zod';
-
-// pause a execution to streamed invocation state with 'call'
-async function sleep(ms) {
-  return new Promise((resolve) => setTimeout(resolve, ms));
-}
-
-const server = new McpServer({
-  name: 'Calc',
-  version: '1.0.0',
-});
-
-server.tool(
-  'calc_add',
-  'Calculate numbers addition, if text has numbers and operation, return the result',
-  { a: z.number(), b: z.number() },
-  async ({ a, b }) => {
-    await sleep(100);
-
-    return {
-      content: [{ type: 'text', text: String(a + b) }],
-    };
-  },
-);
-
-server.tool(
-  'calc_minus',
-  'Calculate numbers subtraction, if text has numbers and operation, return the result',
-  { a: z.number(), b: z.number() },
-  async ({ a, b }) => {
-    await sleep(100);
-
-    return {
-      content: [{ type: 'text', text: String(a - b) }],
-    };
-  },
-);
-
-server.tool(
-  'calc_multiply',
-  'Calculate numbers multiplication, if text has numbers and operation, return the result',
-  { a: z.number(), b: z.number() },
-  async ({ a, b }) => {
-    await sleep(100);
-
-    return {
-      content: [{ type: 'text', text: String(a * b) }],
-    };
-  },
-);
-
-server.tool(
-  'calc_divide',
-  'Calculate numbers division, if text has numbers and operation, return the result',
-  { a: z.number(), b: z.number() },
-  async ({ a, b }) => {
-    await sleep(100);
-
-    return {
-      content: [{ type: 'text', text: b === 0 ? 0 : String(a / b) }],
-    };
-  },
-);
-
-async function main() {
-  const transport = new StdioServerTransport();
-  await server.connect(transport);
-}
-
-main().catch((error) => {
-  console.error(error);
-  process.exit(1);
-});
